Allow opening a Trident/IE tab...
using the current page from the File menu when using GeckoFX (or Trident, doesn't matter).

Currently, the Trident tab gets its URL from the address bar's current text, so that needs to update when the browser loads and stuff.

The "Trident Tab" (basically just IETab, but a bit more integrated sorta like Chromium Edge's IE Tab feature) should have the option to put text in the statusbar to say that the browser is in "Trident Tab" mode.